requirements.txt
gradio
torch
torchvision
pillow
huggingface_hub